requirements.txt
python-dotenv>=1.0.0
streamlit>=1.31.0
mistralai>=0.1.0
PyMuPDF>=1.23.0
docling>=1.0.0
faiss-cpu>=1.7.4 # For local vector store
numpy>=1.24.0
pandas>=2.0.0
pytest>=7.0.0
pydantic>=2.0.0
langchain-text-splitters>=0.2.0
SQLAlchemy>=2.0.0
bcrypt>=4.0.0
plotly>=5.14.0
